watcher/pyproject.toml
[tool.poetry]
name = "watcher"
version = "1.0.0"
description = "Pulls releases of kubectl"
authors = ["Ahmetcan Güvendiren <ahmtcngvndrn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.54"
requests = ">=2.14.0,<2.25="
semver = "^2.13.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
